Rename all populate_obs_*_occ.py to populate_obs_*_prof.prof.py
